[AC-5096] refactor functional_expertise_helper and industry_helper:
- create superclass mptt_model_helper.
- derive functional_expertise_helper and industry_helper from it.
- reference shared CONSTANTS instead of duplicated.
