+++ web/impact/impact/v1/helpers/__init__.py
@@ -3,12 +3,13 @@
 
 
 from impact.v1.helpers.functional_expertise_helper import (
-    FUNCTIONAL_EXPERTISE_FIELDS,
     FunctionalExpertiseHelper,
 )
-from impact.v1.helpers.industry_helper import (
-    INDUSTRY_FIELDS,
-    IndustryHelper,
+from impact.v1.helpers.industry_helper import IndustryHelper
+from impact.v1.helpers.mptt_model_helper import (
+    MPTT_FIELDS,
+    MPTT_TYPE,
+    MPTTModelHelper,
 )
 from impact.v1.helpers.model_helper import (
     BOOLEAN_FIELD,
