+++ web/impact/impact/v1/helpers/functional_expertise_helper.py
@@ -2,37 +2,8 @@
 # Copyright (c) 2017 MassChallenge, Inc.
 
 from impact.models import FunctionalExpertise
-from impact.v1.helpers.model_helper import (
-    ModelHelper,
-    PK_FIELD,
-    REQUIRED_STRING_FIELD,
-    READ_ONLY_ID_FIELD,
-    READ_ONLY_STRING_FIELD,
-    json_object,
-    properties_from_fields,
-)
-
-FUNCTIONAL_EXPERTISE_FIELDS = {
-    "id": PK_FIELD,
-    "name": REQUIRED_STRING_FIELD,
-    "full_name": READ_ONLY_STRING_FIELD,
-    "parent_id": READ_ONLY_ID_FIELD,
-}
-FUNCTIONAL_EXPERTISE_TYPE = json_object(
-    properties_from_fields(FUNCTIONAL_EXPERTISE_FIELDS))
+from impact.v1.helpers.mptt_model_helper import MPTTModelHelper
 
 
-class FunctionalExpertiseHelper(ModelHelper):
+class FunctionalExpertiseHelper(MPTTModelHelper):
     model = FunctionalExpertise
-
-    REQUIRED_KEYS = ["name"]
-    OPTIONAL_KEYS = ["parent_id"]
-    INPUT_KEYS = REQUIRED_KEYS + OPTIONAL_KEYS
-
-    @property
-    def full_name(self):
-        return str(self.subject)
-
-    @classmethod
-    def fields(cls):
-        return FUNCTIONAL_EXPERTISE_FIELDS
